% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condiments-package.R
\docType{package}
\name{condiments-package}
\alias{condiments}
\alias{condiments-package}
\title{condiments: Differential Topology, Progression and Differentiation}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package encapsulate many functions to conduct a differential topology analysis. It focuses on analyzing an 'omic dataset with multiple conditions. While the package is mostly geared toward scRNASeq, it does not place any restriction on the actual input format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hectorrdb.github.io/condiments/index.html}
  \item Report bugs at \url{https://github.com/HectorRDB/condiments/issues}
}

}
\author{
\strong{Maintainer}: Hector Roux de Bezieux \email{hector.rouxdebezieux@berkeley.edu} (\href{https://orcid.org/0000-0002-1489-8339}{ORCID})

Authors:
\itemize{
  \item Koen Van den Berge [contributor]
  \item Kelly Street [contributor]
}

}
\keyword{internal}
