% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiationTest.R
\name{differentiationTest}
\alias{differentiationTest}
\title{Differential differentiation}
\usage{
differentiationTest(...)
}
\arguments{
\item{...}{See the \code{\link{fateSelectionTest}}}
}
\value{
See the \code{\link{fateSelectionTest}}
}
\description{
Test whether or not the cell repartition between lineages is
independent of the conditions
}
\examples{
data('slingshotExample', package = "slingshot")
rd <- slingshotExample$rd
cl <- slingshotExample$cl
condition <- factor(rep(c('A','B'), length.out = nrow(rd)))
condition[110:139] <- 'A'
sds <- slingshot::slingshot(rd, cl)
differentiationTest(sds, condition)
}
