% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffDifferentiationTest.R,
%   R/diffTopoTest.R
\name{diffTopoTest}
\alias{diffTopoTest}
\alias{diffProgressionTest,SingleCellExperiment-method}
\alias{diffTopoTest,SlingshotDataSet-method}
\alias{diffTopoTest,SingleCellExperiment-method}
\title{Differential Topology Test}
\usage{
diffTopoTest(sds, ...)

\S4method{diffProgressionTest}{SingleCellExperiment}(
  sds,
  cd,
  global = TRUE,
  lineages = FALSE,
  method = "KS",
  thresh = 0.05
)

\S4method{diffTopoTest}{SlingshotDataSet}(sds, cd, rep = 200, thresh = 0.05)

\S4method{diffTopoTest}{SingleCellExperiment}(sds, cd = cd, rep = 200, thresh = 0.05)
}
\arguments{
\item{sds}{A slingshot object already run on the full dataset. Can be either a
\code{\link{slingshotDataset}} or a \code{\link{SingleCellExperiment}} object.}

\item{cd}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector.}

\item{thresh}{the threshold for the KS test. See \code{\link{ks_test}}.}

\item{rep}{How many permutations to run. Default to 200.}
}
\description{
Test whether or not slingshot should be fitted independently
for different conditions or not.
}
\examples{
data('slingshotExample')
rd <- slingshotExample$rd
cl <- slingshotExample$cl
condition <- factor(rep(c('A','B'), length.out = nrow(rd)))
condition[110:139] <- 'A'
sds <- getLineages(rd, cl)
diffTopoTest(sds, condition, rep = 20)
}
