% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/proximity_score.R
\name{proximity_score}
\alias{proximity_score}
\alias{proximity_score,matrix-method}
\alias{proximity_score,SingleCellExperiment-method}
\title{Proximity Score}
\usage{
proximity_score(Object, ...)

\S4method{proximity_score}{matrix}(Object, conditions, k = 10, smooth = 10)

\S4method{proximity_score}{SingleCellExperiment}(Object, dimred = 1, conditions, k = 10, smooth = 10)
}
\arguments{
\item{Object}{A \code{\link{SingleCellExperiment}} object or a matrix
representing the reduced dimension matrix of the cells.}

\item{...}{parameters including:}

\item{conditions}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector}

\item{k}{The number of neighbors to consider when computing the score.
Default to 10.}

\item{smooth}{The smoothing parameter. Default to k. Lower values mean that
we smooth more.}

\item{dimred}{A string or integer scalar indicating the reduced dimension
result in \code{reducedDims(sce)} to plot. Default to 1.}
}
\value{
Either a list with the \code{scaled_scores} and the \code{scores} for
 each cell, if input is a matrix, or the \code{\link{SingleCellExperiment}}
 object, wit this list in the \code{\link{colData}}.
}
\description{
Compute a proximity score to show whether nearby cells have the
same condition of not
}
\examples{
sd <- create_differential_topology(n_cells = 200, shift = 0,
                                   unbalance_level = 1)
scores <- proximity_score(sd$rd, sd$conditions, k = 4)
cols <- as.numeric(cut(scores$scaled_scores, 8))
plot(sd$rd[, "Dim1"], sd$rd[, "Dim2"], xlab = "Dim1", ylab = "Dim2",
 pch = 16, col = RColorBrewer::brewer.pal(8, "Blues")[cols])
}
