% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{proximityScore}
\alias{proximityScore}
\title{Proximity Score}
\usage{
proximity_score(Object, ...)
}
\arguments{
\item{k}{The number of neighbours to consider when computing the score.
Default to 10.}

\item{smooth}{The smoothing parameter. Default to k. Lower values mean that
we smooth more.}
}
\description{
Proximity Score
}
