% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/multiple_samples.R
\name{topologyTest_multipleSamples}
\alias{topologyTest_multipleSamples}
\alias{topologyTest_multipleSamples,SingleCellExperiment-method}
\alias{topologyTest_multipleSamples,PseudotimeOrdering-method}
\title{Differential Topology Test with multiple samples}
\usage{
topologyTest_multipleSamples(sds, ...)

\S4method{topologyTest_multipleSamples}{SingleCellExperiment}(sds, conditions, Samples, ...)

\S4method{topologyTest_multipleSamples}{PseudotimeOrdering}(sds, conditions, Samples, ...)
}
\arguments{
\item{...}{parameters including:}
}
\description{
Differential Topology Test with multiple samples
}
