% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_differential_topology}
\alias{create_differential_topology}
\title{Create Example function}
\usage{
create_differential_topology(
  n_cells = 200,
  noise = 0.15,
  shift = 10,
  unbalance_level = 0.9
)
}
\arguments{
\item{n_cells}{The number of cells in the dataset.}

\item{noise}{Amount of noise. Between 0 and 1.}

\item{shift}{How much should the top lineage shift in condition B.}

\item{unbalance_levelHow}{much should the bottom lineage be unbalanced toward
condition A.}
}
\value{
A list with three components
 \itemize{
  \item \code{rd}: The reduced dimensions coordinates of every cells. An
  \code{n_cells} by \code{2} matrix
  \item \code{lineages}: A vector of length \code{n_cells}. Either 1 or 2
  depending of lineage assignment for each cell.
  \item \code{cd}: A vector of length \code{n_cells}. Either A or B
  depending of condition assignment for each cell.
}
}
\description{
This creates a simulated reduced dimension dataset
}
\examples{
sd <- create_differential_topology()
}
