% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffDifferentiationTest.R
\name{diffDifferentiationTest}
\alias{diffDifferentiationTest}
\alias{diffDifferentiationTest,SlingshotDataSet-method}
\title{Differential Differentiation Test}
\usage{
diffDifferentiationTest(sds, ...)

\S4method{diffDifferentiationTest}{SlingshotDataSet}(
  sds,
  cd,
  global = TRUE,
  lineages = FALSE,
  method = "KS",
  thresh = 0.05
)
}
\arguments{
\item{sds}{The final object after running slingshot. Can be either a
\code{\link{slingshotDataset}} or a \code{\link{SingleCellExperiment}} object.}

\item{cd}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector}

\item{global}{If TRUE, test for all pairs simultaneously.}

\item{method}{Either "KS" for the weighted Kolmogorov-Smirnov or "Permutation"
for a permutation. See details. Default to KS.}

\item{thresh}{the threshold for the KS test. See \code{\link{ks_test}}.}

\item{pairwise}{If TRUE, test for all pairs independently.}
}
\description{
Test whether or not the cell repartition between lineages is
independent of the conditions
}
