% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffDifferentiationTest.R
\name{diffDifferentiationTest}
\alias{diffDifferentiationTest}
\alias{diffDifferentiationTest,SlingshotDataSet-method}
\title{Differential Differentiation Test}
\usage{
diffDifferentiationTest(sds, ...)

\S4method{diffDifferentiationTest}{SlingshotDataSet}(
  sds,
  conditions,
  global = TRUE,
  lineages = FALSE,
  method = "Permutation"
)
}
\arguments{
\item{sds}{The final object after running slingshot. Can be either a
\code{\link{slingshotDataset}} or a \code{\link{SingleCellExperiment}} object.}

\item{conditions}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector}

\item{global}{If TRUE, test for all pairs simultaneously.}

\item{method}{For now, only "Permutation" is accepted.}

\item{pairwise}{If TRUE, test for all pairs independently.}
}
\description{
Test whether or not the cell repartition between lineages is
independent of the conditions
}
