% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffDifferentiationTest.R
\name{diffDifferentiationTest}
\alias{diffDifferentiationTest}
\alias{diffDifferentiationTest,SlingshotDataSet-method}
\alias{diffDifferentiationTest,SingleCellExperiment-method}
\title{Differential Differentiation Test}
\usage{
diffDifferentiationTest(sds, ...)

\S4method{diffDifferentiationTest}{SlingshotDataSet}(
  sds,
  conditions,
  global = TRUE,
  pairwise = FALSE,
  method = "Permutation"
)

\S4method{diffDifferentiationTest}{SingleCellExperiment}(
  sds,
  conditions,
  global = TRUE,
  pairwise = FALSE,
  method = "Permutation"
)
}
\arguments{
\item{sds}{The final object after running slingshot. Can be either a
\code{\link{SlingshotDataSet}} or a \code{\link{SingleCellExperiment}} object.}

\item{...}{parameters including:}

\item{conditions}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector}

\item{global}{If TRUE, test for all pairs simultaneously.}

\item{pairwise}{If TRUE, test for all pairs independently.}

\item{method}{For now, only "Permutation" is accepted.}
}
\value{
A data frame with 3 columns:
\itemize{
  \item *Pval* the pvalue for the test at the global or lineage level
  \item *Statistic* for individual lineages, either the modified KS statistic
  if \code{method = "KS"}, or the weighted difference of means, if
  \code{method = "Permutation"}. For the global test, the combined Z-score.
  \item *Pair* for individual pairs, the lineages numbers. For global,
  \code{"All"}.
}
}
\description{
Test whether or not the cell repartition between lineages is
independent of the conditions
}
\examples{
data('slingshotExample', package = "slingshot")
rd <- slingshotExample$rd
cl <- slingshotExample$cl
condition <- factor(rep(c('A','B'), length.out = nrow(rd)))
condition[110:139] <- 'A'
sds <- slingshot::slingshot(rd, cl)
diffDifferentiationTest(sds, condition)
}
