% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffDifferentiationTest.R
\name{diffDifferentiationTest}
\alias{diffDifferentiationTest}
\alias{diffDifferentiationTest,SlingshotDataSet-method}
\alias{diffDifferentiationTest,SingleCellExperiment-method}
\title{Differential Differentiation Test}
\usage{
diffDifferentiationTest(sds, ...)

\S4method{diffDifferentiationTest}{SlingshotDataSet}(
  sds,
  conditions,
  global = TRUE,
  pairwise = FALSE,
  method = "Classifier",
  thresh = 0.05,
  ...
)

\S4method{diffDifferentiationTest}{SingleCellExperiment}(
  sds,
  conditions,
  global = TRUE,
  pairwise = FALSE,
  method = "Classifier",
  thresh = 0.05,
  ...
)
}
\arguments{
\item{sds}{The final object after running slingshot. Can be either a
\code{\link{SlingshotDataSet}} or a \code{\link{SingleCellExperiment}} object.}

\item{...}{Other arguments passed to \link[Ecume]{classifier_test}.}

\item{conditions}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector}

\item{global}{If TRUE, test for all pairs simultaneously.}

\item{pairwise}{If TRUE, test for all pairs independently.}

\item{method}{For now, only "Classifier" is accepted.}

\item{thresh}{The threshold for the classifier test. See details.
Default to .05.}
}
\value{
A data frame with 3 columns:
\itemize{
  \item *pair* for individual pairs, the lineages numbers. For global,
  \code{"All"}.
  \item *p.value* the pvalue for the test at the global or pair level
  \item *statistic* The classifier accuracy
}
}
\description{
Test whether or not the cell repartition between lineages is
independent of the conditions
}
\examples{
data('slingshotExample', package = "slingshot")
rd <- slingshotExample$rd
cl <- slingshotExample$cl
condition <- factor(rep(c('A','B'), length.out = nrow(rd)))
condition[110:139] <- 'A'
sds <- slingshot::slingshot(rd, cl)
diffDifferentiationTest(sds, condition)
}
