% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/differentiationTest.R
\name{differentiationTest}
\alias{differentiationTest}
\alias{differentiationTest,matrix-method}
\alias{differentiationTest,SlingshotDataSet-method}
\alias{differentiationTest,SingleCellExperiment-method}
\title{Differential Differentiation Test}
\usage{
differentiationTest(cellWeights, ...)

\S4method{differentiationTest}{matrix}(
  cellWeights,
  conditions,
  global = TRUE,
  pairwise = FALSE,
  method = c("mmd", "Classifier"),
  thresh = 0.05,
  ...
)

\S4method{differentiationTest}{SlingshotDataSet}(
  cellWeights,
  conditions,
  global = TRUE,
  pairwise = FALSE,
  method = c("mmd", "Classifier"),
  thresh = 0.05,
  ...
)

\S4method{differentiationTest}{SingleCellExperiment}(
  cellWeights,
  conditions,
  global = TRUE,
  pairwise = FALSE,
  method = c("mmd", "Classifier"),
  thresh = 0.05,
  ...
)
}
\arguments{
\item{cellWeights}{Can be either a \code{\link{SlingshotDataSet}}, a
\code{\link{SingleCellExperiment}} object or a matrix of cell weights
defining the probability that a cell belongs to a particular lineage.
Each row represents a cell and each column represents a lineage. If only a
single lineage, provide a matrix with one column containing all values of 1.}

\item{...}{Other arguments passed to \link[Ecume]{classifier_test}.}

\item{conditions}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector}

\item{global}{If TRUE, test for all pairs simultaneously.}

\item{pairwise}{If TRUE, test for all pairs independently.}

\item{method}{One of "Classifier" or "mmd".}

\item{thresh}{The threshold for the classifier test. See details.
Default to .05.}
}
\value{
A data frame with 3 columns:
\itemize{
  \item *pair* for individual pairs, the lineages numbers. For global,
  \code{"All"}.
  \item *p.value* the pvalue for the test at the global or pair level
  \item *statistic* The classifier accuracy
}
}
\description{
Test whether or not the cell repartition between lineages is
independent of the conditions
}
\examples{
data('slingshotExample', package = "slingshot")
rd <- slingshotExample$rd
cl <- slingshotExample$cl
condition <- factor(rep(c('A','B'), length.out = nrow(rd)))
condition[110:139] <- 'A'
sds <- slingshot::slingshot(rd, cl)
differentiationTest(sds, condition)
}
