% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffProgressionTest.R
\name{diffProgressionTest}
\alias{diffProgressionTest}
\alias{diffProgressionTest,SlingshotDataSet-method}
\alias{diffProgressionTest,SingleCellExperiment-method}
\title{Differential Progression Test}
\usage{
diffProgressionTest(sds, ...)

\S4method{diffProgressionTest}{SlingshotDataSet}(
  sds,
  cd,
  global = TRUE,
  lineages = FALSE,
  method = "KS",
  thresh = 0.05
)

\S4method{diffProgressionTest}{SingleCellExperiment}(
  sds,
  cd,
  global = TRUE,
  lineages = FALSE,
  method = "KS",
  thresh = 0.05
)
}
\arguments{
\item{sds}{The final object after running slingshot. Can be either a
\code{\link{slingshotDataset}} or a \code{\link{SingleCellExperiment}} object.}

\item{cd}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector.}

\item{global}{If TRUE, test for all lineages simultaneously.}

\item{lineages}{If TRUE, test for all lineages independently.}

\item{method}{Either "KS" for the weighted Kolmogorov-Smirnov or "Permutation"
for a permutation. See details. Default to KS.}

\item{thresh}{the threshold for the KS test. See \code{\link{ks_test}}.}
}
\description{
Test whether or not the pseudotime distribution are identical
within lineages between conditions
}
