% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffProgressionTest.R,
%   R/diffTopoTest.R
\name{diffTopoTest}
\alias{diffTopoTest}
\alias{diffProgressionTest,SlingshotDataSet-method}
\alias{diffProgressionTest,SingleCellExperiment-method}
\alias{diffTopoTest,SlingshotDataSet-method}
\alias{diffTopoTest,SingleCellExperiment-method}
\title{Differential Topology Test}
\usage{
diffTopoTest(sds, ...)

diffProgressionTest(sds, ...)

\S4method{diffProgressionTest}{SlingshotDataSet}(sds, thresh = 0.05, global = TRUE, lineages = FALSE)

\S4method{diffProgressionTest}{SingleCellExperiment}(sds, rep = 200, thresh = 0.05)

\S4method{diffTopoTest}{SlingshotDataSet}(sds, rep = 200, thresh = 0.05)

\S4method{diffTopoTest}{SingleCellExperiment}(sds, rep = 200, thresh = 0.05)
}
\arguments{
\item{sds}{A slingshot object already run on the full dataset. Can be either a
\code\link{{slingshotDataset}} or a \code\link{{SingleCellExperiment}} object.}

\item{thresh}{the threshold for the KS test. See \code{\link{ks_test}}.}

\item{global}{If TRUE, test for all lineages simultaneously.}

\item{lineages}{If TRUE, test for all lineages independently.}

\item{rep}{How many permutations to run. Default to 200.}
}
\description{
Test whether or not slingshot should be fitted independently
for different conditions or not.

Test whether or not slingshot should be fitted independently
for different conditions or not.
}
\examples{
sd <- create_differential_topology(n_cells = 200, shift = 0,
                                   unbalance_level = 1)
sd <- create_differential_topology(n_cells = 200, shift = 0,
                                   unbalance_level = 1)
}
