% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/diffTopoTest.R
\name{diffTopoTest}
\alias{diffTopoTest}
\alias{diffTopoTest,SlingshotDataSet-method}
\alias{diffTopoTest,SingleCellExperiment-method}
\title{Differential Topology Test}
\usage{
diffTopoTest(sds, ...)

\S4method{diffTopoTest}{SlingshotDataSet}(
  sds,
  conditions,
  rep = 200,
  thresh = 0.05,
  method = ifelse(dplyr::n_distinct(conditions) == 2, "KS_mean", "Classifier"),
  ...
)

\S4method{diffTopoTest}{SingleCellExperiment}(
  sds,
  conditions,
  rep = 200,
  thresh = 0.05,
  method = ifelse(dplyr::n_distinct(conditions) == 2, "KS_mean", "Classifier"),
  ...
)
}
\arguments{
\item{sds}{A slingshot object already run on the full dataset. Can be either a
\code{\link{SlingshotDataSet}} or a \code{\link{SingleCellExperiment}} object.}

\item{...}{Other arguments passed to the test method. See details}

\item{conditions}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector.}

\item{rep}{How many permutations to run. Default to 50.}

\item{thresh}{the threshold for the KS test. See \code{\link{ks_test}}.}

\item{method}{The method to use to test. One of 'KS_mean', "KS_all', "mmd'
and 'Classifier'. See details.}
}
\value{
A list containing the following components:
\itemize{
  \item *statistic* the value of the test statistic.
  \item *p.value* the p-value of the test.
}
}
\description{
Test whether or not slingshot should be fitted independently
for different conditions or not.
}
\details{
If there is only two conditions, default to `KS_mean`. Otherwise,
uses a classifier.
}
\examples{
data('slingshotExample', package = "slingshot")
rd <- slingshotExample$rd
cl <- slingshotExample$cl
condition <- factor(rep(c('A','B'), length.out = nrow(rd)))
condition[110:139] <- 'A'
sds <- slingshot::getLineages(rd, cl)
diffTopoTest(sds, condition, rep = 20)
}
