% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_k_test.R
\name{ks_test}
\alias{ks_test}
\title{New KS Test}
\usage{
ks_test(x, y, thresh = 0.05)
}
\arguments{
\item{x}{Vector of values sampled from the first distribution}

\item{y}{Vector of values sampled from the second distribution}

\item{thresh}{The threshold needed to clear between the two cumulative distributions}
}
\description{
Kolmogorov-Smirnov Two-Sample Test, with threshold for differences
}
\examples{
 x <- runif(100)
 y <- runif(100, min = .001, max = .001)
 .ks.test(x, y, thresh = .001)
}
