% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nLineages.R
\name{nLineages}
\alias{nLineages}
\alias{nLineages,SingleCellExperiment-method}
\alias{nLineages,SlingshotDataSet-method}
\title{Number of lineages}
\usage{
nLineages(sds, ...)

\S4method{nLineages}{SingleCellExperiment}(sds)

\S4method{nLineages}{SlingshotDataSet}(sds)
}
\arguments{
\item{...}{parameters including:}
}
\description{
Return the number of lineages for a slingshot object
}
