% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/topologyTest.R
\name{topologyTest}
\alias{topologyTest}
\alias{topologyTest,SlingshotDataSet-method}
\alias{topologyTest,SingleCellExperiment-method}
\title{Differential Topology Test}
\usage{
topologyTest(sds, ...)

\S4method{topologyTest}{SlingshotDataSet}(
  sds,
  conditions,
  rep = 200,
  threshs = 0.05,
  methods = ifelse(dplyr::n_distinct(conditions) == 2, "KS_mean", "Classifier"),
  args_mmd = list(),
  args_classifier = list()
)

\S4method{topologyTest}{SingleCellExperiment}(
  sds,
  conditions,
  rep = 200,
  threshs = 0.05,
  methods = ifelse(dplyr::n_distinct(conditions) == 2, "KS_mean", "Classifier"),
  args_mmd = list(),
  args_classifier = list()
)
}
\arguments{
\item{sds}{A slingshot object already run on the full dataset. Can be either a
\code{\link{SlingshotDataSet}} or a \code{\link{SingleCellExperiment}} object.}

\item{...}{parameters including:}

\item{conditions}{Either the vector of conditions, or a character indicating which
column of the metadata contains this vector.}

\item{rep}{How many permutations to run. Default to 50.}

\item{threshs}{the threshold(s) for the KS test or classifier test.
See \code{\link{ks_test}} and \code{\link{classifier_test}}.}

\item{methods}{The method(s) to use to test. Must be among 'KS_mean',
"KS_all', "mmd', and 'Classifier'. See details.}

\item{args_mmd}{arguments passed to the mmd test. See \code{\link{mmd_test}}.}

\item{args_classifier}{arguments passed to the classifier test. See \code{\link{classifier_test}}.}
}
\value{
A list containing the following components:
\itemize{
  \item *method* The method used to test
  \item *thresh* The threshold (if relevant)
  \item *statistic* the value of the test statistic.
  \item *p.value* the p-value of the test.
}
}
\description{
Test whether or not slingshot should be fitted independently
for different conditions or not.
}
\details{
If there is only two conditions, default to `KS_mean`. Otherwise,
uses a classifier.
}
\examples{
data('slingshotExample', package = "slingshot")
rd <- slingshotExample$rd
cl <- slingshotExample$cl
condition <- factor(rep(c('A','B'), length.out = nrow(rd)))
condition[110:139] <- 'A'
sds <- slingshot::getLineages(rd, cl)
topologyTest(sds, condition, rep = 20)
}
